# swagger.yaml
openapi: 3.0.3
info:
  title: TaskMan API
  version: 1.3.0
  description: >
    REST API for project and task management with JWT authentication,
    Google OAuth2 login and role-based access control (RBAC).
servers:
  - url: http://localhost:8080/api
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    TokenPair:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
...
paths:
...
  /auth/register:
    post:
      tags: [Auth]
      summary: Register with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, email, password]
              properties:
                name:
                  type: string
                email:
                  type: string
                password:
                  type: string
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenPair"
        "409":
          description: Email already registered
  /auth/login:
    post:
      tags: [Auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                password:
                  type: string
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenPair"
        "401":
          description: Invalid credentials
...
  /auth/refresh-token:
    post:
      tags: [Auth]
      summary: Refresh JWT tokens
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken:
                  type: string
      responses:
        "200":
          description: New token pair
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenPair"

  # =====================
  # USERS
  # =====================
...